This would require more work
The KeyValue interface is too basic. It doesn't provide any sort of
context about what could be coming next. rawdb module just hands out
parts of the block without any clear signal about the boundaries.

To make this work, it would require that we absorb the parts as soon as
they're given and later try to consolidate those parts into a block as
was originally seen by the rawdb module. Or, somehow try to determine
the intentions of rawdb module.

This whole thing could be a lot simpler, if rawdb module itself could
provide a better interface -- an interface which can hand out bigger
chunks of data instead of each part separately -- sort of what they've
done for the ancient store.

Ultimately, this isn't as straightforward as implementing the key-value
interface. I don't feel like continuing here, unless the ETH foundation
decides to fund the effort.